refactor(zip): Factor zip package out of compartment mapper
