+++ package.json
@@ -16,7 +16,8 @@
     "packages/ses-ava",
     "packages/ses-integration-test",
     "packages/test262-runner",
-    "packages/transform-module"
+    "packages/transform-module",
+    "packages/zip"
   ],
   "engines": {
     "node": ">=12.17.0"
