+++ packages/compartment-mapper/package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "@endo/lexer": "^0.1.0",
     "@endo/static-module-record": "^0.1.0",
+    "@endo/zip": "^0.1.0",
     "ses": "^0.12.7"
   },
   "devDependencies": {
