+++ packages/compartment-mapper/src/archive.js
@@ -1,7 +1,7 @@
 // @ts-check
 /* eslint no-shadow: 0 */
 
-import { writeZip } from './zip.js';
+import { writeZip } from '@endo/zip';
 import { resolve } from './node-module-specifier.js';
 import { compartmentMapForNodeModules } from './node-modules.js';
 import { search } from './search.js';
