+++ packages/zip/index.d.ts
@@ -1,0 +1,3 @@
+export { readZip, writeZip } from './src/main.js';
+export { ZipReader } from './src/reader.js';
+export { ZipWriter } from './src/writer.js';
