+++ packages/zip/src/crc32.js
