+++ packages/zip/src/exported.js
@@ -1,0 +1,28 @@
+/**
+ * @typedef {Object} ArchiveReader
+ * @property {ReadFn} read
+ */
+
+/**
+ * @callback ReadFn
+ * @param {string} name
+ * @returns {Promise<Uint8Array>} bytes
+ */
+
+/**
+ * @typedef {Object} ArchiveWriter
+ * @property {WriteFn} write
+ * @property {SnapshotFn} snapshot
+ */
+
+/**
+ * @callback WriteFn
+ * @param {string} name
+ * @param {Uint8Array} bytes
+ * @returns {Promise<void>}
+ */
+
+/**
+ * @callback SnapshotFn
+ * @returns {Promise<Uint8Array>}
+ */
