+++ packages/zip/src/reader.js
