+++ packages/zip/src/signature.js
